requirements.txt
pandas
numpy
sqlalchemy
openpyxl
python-dotenv
scipy
